Instead of getting it to work on 10.3.9, I turned off Panther support. :-(
Additionally, I finally fixed the warnings the XADMaster project was giving us and made some other changes.

+++ PGActivityPanelController.m
@@ -70,13 +70,13 @@ DEALINGS WITH THE SOFTWARE. */
 {
 	return item ? [[item subloads] count] > 0 : YES;
 }
-- (NSInteger)outlineView:(NSOutlineView *)outlineView
-             numberOfChildrenOfItem:(id)item
+- (int)outlineView:(NSOutlineView *)outlineView
+       numberOfChildrenOfItem:(id)item
 {
 	return [[(item ? item : [PGLoadManager sharedLoadManager]) subloads] count];
 }
 - (id)outlineView:(NSOutlineView *)outlineView
-      child:(NSInteger)index
+      child:(int)index
       ofItem:(id)item
 {
 	return [[(item ? item : [PGLoadManager sharedLoadManager]) subloads] objectAtIndex:index];
